+++ site/api/email.lua
@@ -83,14 +83,7 @@ function handle(r)
                 end
             -- Or do we just want the email itself?
             else
-                doc.from = doc.from or "unknown"
-                local eml = doc.from:match("<(.-)>") or doc.from:match("%S+@%S+") or nil
-                if eml == nil and doc.from:match(".- at .- %(") then
-                    eml = doc.from:match("(.- at .-) %("):gsub(" at ", "@")
-                    doc.from = eml
-                elseif eml == nil then
-                    eml = "unknown"
-                end
+                local eml = utils.extractCanonEmail(doc.from or "unknown")
                 if not account then -- anonymize email address if not logged in
                     doc = utils.anonymizeHdrs(doc, true)
                 end